Avoid concurrent rollback from rollbacker and rollback fragment
This patch avoids the situation where rollbacker is fired and
a rollback fragment attempt to rollback the same transaction
concurrently. This happens when a high priority streaming transaction
is BF aborted by TOI, and at the same time the transaction managed to
replicate one more fragment which fails certification. Failed
certification results in rollback fragment, which may be delivered
and processes before the streaming applier is stopped by in bf_abort().
The fix consists in changing rollback fragment processing so that
it stops the corresponding streaming applier right away.
Should a rollback fragment and BF abort happen concurrently, we now
expect that only one of the two manages stop the affected streaming
applier. So that the transaction is either rolled back by rollback
fragment or rollbacker thread in DBMS.