Enable DBMS initialization from applying codepath
In general the provider might not guarantee that after joining
to replication group, the first applied event is view event.

Made init_initialized_ atomic and check it also from applying
codepath. If the DBMS is not yet initialized when applying the
first event, change state to initializing and wait for DBMS
initialization completion.

Notes:
- Allow connected -> initializing state change.
- Handle bootstrap_view in mock_connect to make server state synced
  for tests.