Improve error handling of transaction::commit_or_rollback_by_xid()
Handle the case where a call to commit_or_rollback_by_xid() is given
a xid for which there is no corresponding streaming applier.
We distinguish two cases here:
1) the xid might not exist at all, or the corresponding transaction
was already committed or rolled back. The client may just return an
error; or 2) all streaming appliers have been closed
because the node is currently disconnected. We can't tell if a
corresponding transaction exists. The client may want to retry.